DHFPROD-4559: Refactoring SearchHelper, SearchManager classes and deprecated content
